Adjstments suggested by aiky30
